# docs/swagger.yaml
basePath: /
...
host: localhost:8080
info:
  contact: {}
  description: Career discovery API for Nepal
  title: Carevo API
  version: "1.0"
paths:
...
  /careers/{id}/resources:
    get:
      parameters:
      - description: Career ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get resources for a career
      tags:
      - careers
  /careers/{id}/roadmap:
    get:
      parameters:
      - description: Career ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get roadmap steps for a career
      tags:
      - careers
  /careers/{slug}:
    get:
      parameters:
      - description: Career slug
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get career by slug
      tags:
      - careers
  /careers/search:
    get:
      parameters:
      - description: Search query
        in: query
        name: q
        required: true
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Search careers
      tags:
      - careers
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"